// fe/src/shared/types/enums/education.ts
export enum EducationType {
  TRAINING = 'TRAINING', // повышение квалификации / курсы
  PRE_MIDDLE = 'PRE_MIDDLE', // неоконченное среднее
  MIDDLE = 'MIDDLE', // среднее
  PRE_HIGHT = 'PRE_HIGHT', // неоконченное высшее
  SPECIALIST = 'SPECIALIST', // специалист
  BACHELOR = 'BACHELOR', // бакалавр
  MASTER = 'MASTER', // магистр
  PRE_PHD = 'PRE_PHD', // кандидат наук
  PHD = 'PHD', // доктор наук
}

export enum EducationOrganizationType {
  UNIVERSITY = 'UNIVERSITY',
  HIGH_SCHOOL = 'HIGH_SCHOOL',
  SCHOOL = 'SCHOOL',
  COURSE = 'COURSE',
}